+++ WarpDrive/events/glcolorevent.hpp
@@ -1,6 +1,8 @@
 #ifndef WD_GL_COLOR_EVENT_HPP_DEFINED
 #define WD_GL_COLOR_EVENT_HPP_DEFINED
 
+#include <cstdint>
+
 #include "events/objectevent.hpp"
 
 class glColorEvent: public GOEvent
@@ -9,13 +11,13 @@ class glColorEvent: public GOEvent
 */
 public:
 
-	glColorEvent(float Time = 0, float Red = 1, float Green = 1, float Blue = 1);
+    glColorEvent(int64_t Time = 0, float Red = 1, float Green = 1, float Blue = 1);
 	~glColorEvent(){}
     void update(GameObject* target) override;
 
 private:
 
-    float time;
+    int64_t time;
     float r;
     float g;
     float b;
