+++ WarpDrive/events/panandtiltevent.hpp
@@ -1,6 +1,8 @@
 #ifndef WD_PAN_AND_TILT_EVENT_HPP_DEFINED
 #define WD_PAN_AND_TILT_EVENT_HPP_DEFINED
 
+#include <cstdint>
+
 #include "events/objectevent.hpp"
 #include "basemaths/vec3.hpp"
 
@@ -9,13 +11,13 @@ class PanAndTiltEvent: public GOEvent
 
 public:
 
-	PanAndTiltEvent(float Time = 0, Vec3f Pan = Vec3f(), Vec3f Tilt = Vec3f());
+    PanAndTiltEvent(int64_t Time = 0, Vec3f Pan = Vec3f(), Vec3f Tilt = Vec3f());
 	void update(GameObject* target);
 
 protected:
 
 	Vec3f pan, tilt, panNudge, tiltNudge;
-	float time;
+    int64_t time;
 	bool overTime;
 
 };
